Gravity Implemented
Currently implemented as a constant downward force as a velocity of
Vector3f(0, -1, 0). It's not being simulated as a force with some
magnitude along a vector with a tail at the PhysicsObject and a head at
some point because this would be great for an orbit simulator, but not
much else.